Make Toss instruction actions use full-width buttons
