+++ frontend/src/payments/components/TossInstructionDialog.vue
@@ -3,7 +3,7 @@ import { computed } from 'vue'
 
 import { useI18nStore } from '@/localization/store'
 import AccountInfoCard from '@/payments/components/AccountInfoCard.vue'
-import AppDialog from '@/shared/components/AppDialog.vue'
+import DialogBase from '@/shared/components/DialogBase.vue'
 import type { TossPaymentInfo } from '@/payments/services/paymentInfoService'
 
 interface Props {
@@ -26,6 +26,7 @@ const i18nStore = useI18nStore()
 const title = computed(() => i18nStore.t('tossInstruction.title'))
 const description = computed(() => i18nStore.t('tossInstruction.description'))
 const copiedLabel = computed(() => i18nStore.t('tossInstruction.copied'))
+const closeLabel = computed(() => i18nStore.t('dialog.close'))
 const countdownLabel = computed(() => {
   if (props.countdown > 0) {
     return i18nStore.t('tossInstruction.countdown').replace(
@@ -54,11 +55,10 @@ const onLaunchNow = () => {
 </script>
 
 <template>
-  <AppDialog
+  <DialogBase
     :visible="props.visible"
     :title="title"
     :description="description"
-    close-alignment="full"
     @close="onClose"
   >
     <div class="flex flex-col gap-4">
@@ -71,13 +71,13 @@ const onLaunchNow = () => {
         :account-no="props.info.accountNo"
         :account-holder="props.info.accountHolder"
       />
-      <div
-        class="flex items-center justify-center"
-      >
+    </div>
+    <template #footer="{ onClose: closeDialog }">
+      <footer class="mt-6 flex flex-col gap-2">
         <button
           v-if="isCountingDown"
           type="button"
-          class="font-semibold text-roadshop-primary cursor-pointer"
+          class="w-full rounded-xl bg-roadshop-primary px-4 py-2 text-sm font-semibold text-white transition hover:bg-roadshop-primary/90"
           @click="onLaunchNow"
         >
           {{ countdownLabel }}
@@ -85,12 +85,19 @@ const onLaunchNow = () => {
         <button
           v-else
           type="button"
-          class="font-semibold text-roadshop-primary underline underline-offset-2"
+          class="w-full rounded-xl bg-roadshop-primary px-4 py-2 text-sm font-semibold text-white transition hover:bg-roadshop-primary/90"
           @click="onReopen"
         >
           {{ reopenLabel }}
         </button>
-      </div>
-    </div>
-  </AppDialog>
+        <button
+          type="button"
+          class="w-full rounded-xl border border-slate-200 px-4 py-2 text-sm font-medium text-slate-500 transition hover:bg-slate-100"
+          @click="closeDialog()"
+        >
+          {{ closeLabel }}
+        </button>
+      </footer>
+    </template>
+  </DialogBase>
 </template>
